+++ super_editor/lib/src/default_editor/document_hardware_keyboard/document_keyboard_actions.dart
@@ -13,7 +13,6 @@ import 'package:super_editor/src/default_editor/paragraph.dart';
 import 'package:super_editor/src/default_editor/text.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
 import 'package:super_editor/src/infrastructure/keyboard.dart';
-import 'package:super_editor/src/infrastructure/key_event_extensions.dart';
 import 'package:super_editor/src/infrastructure/text_input.dart';
 
 /// Scrolls up by the viewport height, or as high as possible,
