+++ super_editor/lib/src/infrastructure/keyboard.dart
@@ -4,8 +4,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter/services.dart';
 
-import 'package:super_editor/src/infrastructure/key_event_extensions.dart';
-
 
 enum ExecutionInstruction {
   /// The handler has no relation to the key event and
