+++ super_editor/test/super_editor/text_entry/tagging/action_tags_test.dart
@@ -4,7 +4,6 @@ import 'package:flutter_test_robots/flutter_test_robots.dart';
 import 'package:flutter_test_runners/flutter_test_runners.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
-import 'package:logging/logging.dart';
 
 import '../../supereditor_test_tools.dart';
 import '../../test_documents.dart';
